refactor(input): move to script setup
