+++ src/config.json
@@ -388,6 +388,7 @@
           "name": "Input",
           "cName": "输入框",
           "desc": "输入框组件",
+          "setup": true,
           "author": "gxx158"
         },
         {
